Move BackendService to correct package
