+++ app/src/main/java/com/rebalance/service/SoundPlayer.kt
@@ -15,8 +15,8 @@ import androidx.compose.runtime.MutableState
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.rebalance.Preferences
-import com.rebalance.PreferencesData
+import com.rebalance.service.Preferences
+import com.rebalance.service.PreferencesData
 
 @Composable
 fun SoundPlayer(
