+++ app/src/main/java/com/rebalance/ui/component/main/DisplayExpenseImage.kt
@@ -17,7 +17,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.asImageBitmap
 import androidx.compose.ui.unit.dp
-import com.rebalance.PreferencesData
+import com.rebalance.service.PreferencesData
 import com.rebalance.backend.service.BackendService
 import com.rebalance.util.alertUser
 import compose.icons.EvaIcons
