+++ app/src/main/java/com/rebalance/ui/component/main/scaffold/TopAppBar.kt
@@ -9,8 +9,8 @@ import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavHostController
-import com.rebalance.Preferences
-import com.rebalance.PreferencesData
+import com.rebalance.service.Preferences
+import com.rebalance.service.PreferencesData
 import com.rebalance.R
 import com.rebalance.activity.AuthenticationActivity
 import com.rebalance.ui.navigation.Routes
