+++ app/src/main/java/com/rebalance/ui/screen/main/SettingsScreen.kt
@@ -13,7 +13,7 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.rebalance.Preferences
+import com.rebalance.service.Preferences
 import com.rebalance.R
 import com.rebalance.service.notification.NotificationService
 import com.rebalance.ui.component.SoundPlayer
